Update to ic-cdk 0.20

+++ src/ic_util/mod.rs
@@ -27,7 +27,7 @@ pub fn get_cycles() -> u64 {
 pub fn get_stable_memory_size() -> u64 {
     #[cfg(target_arch = "wasm32")]
     {
-        ic_cdk::stable::stable_size() * WASM_PAGE_SIZE
+        (ic_cdk::stable::stable_size() as u64) * WASM_PAGE_SIZE
     }
     #[cfg(not(target_arch = "wasm32"))]
     {
